requirements.txt
gevent>=1.0.2
django>=1.10.0
leancloud-sdk>=1.0.9
Pillow >= 4.0.0
